+++ src/method/linearindex.ts
@@ -8,34 +8,39 @@ import {
     bitapSearch,
     createBitapKey,
 } from "@src/util/algorithm";
+import * as v from "valibot";
 
-type ContentRange = {
-    id: number;
-    path: string;
-    start: number;
-    end: number;
-};
+const ContentRange_v = v.object({
+    id: v.number(),
+    path: v.string(),
+    start: v.number(),
+    end: v.number(),
+});
+type ContentRange = v.InferOutput<typeof ContentRange_v>;
 
-export type LinearIndexEntry = {
-    key: Record<string, unknown>[];
-    content: string;
-    content_length: number;
-    num_id: number;
-    toc: ContentRange[];
-};
+export const LinearIndexEntry_v = v.object({
+    key: v.array(v.record(v.string(), v.unknown())),
+    content: v.string(),
+    content_length: v.number(),
+    num_id: v.number(),
+    toc: v.array(ContentRange_v),
+});
+export type LinearIndexEntry = v.InferOutput<typeof LinearIndexEntry_v>;
 
 export class LinearIndex implements SearchIndex<LinearIndexEntry> {
     public index_entry: LinearIndexEntry;
     public u32_content: Uint32Array;
 
     public constructor(index?: LinearIndexEntry) {
-        this.index_entry = index || {
-            key: [],
-            content: "",
-            content_length: 0,
-            num_id: 0,
-            toc: [],
-        };
+        this.index_entry = index
+            ? v.parse(LinearIndexEntry_v, index)
+            : {
+                  key: [],
+                  content: "",
+                  content_length: 0,
+                  num_id: 0,
+                  toc: [],
+              };
         if (index) {
             this.u32_content = new Uint32Array(this.index_entry.content_length);
             for (let i = 0; i < this.index_entry.content_length; i++) {